fibo

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 /**
  * main - program 50 fibonacci starts here
- * @n
+ * and
  * Return: returns 0 (success)
 */
 int main(void)
 {
-		long long num1, num2, newnum;
+		double num1, num2, newnum;
 		int count;
 
 		num1 = 1;
 		num2 = 2;
-		printf("%lld, %lld, ", num1, num2);
+		printf("%g, %g, ", num1, num2);
 		for (count = 1; count <= 96; count++)
 		{
 		newnum = num1 + num2;
 		if (count < 96)
-		printf("%lld, ", newnum);
+		printf("%g, ", newnum);
 		else
-		printf("%lld\n", newnum);
+		printf("%g\n", newnum);
 		num1 = num2;
 		num2 = newnum;
 		}
